requirements.txt
streamlit>=1.28.1
google-genai>=0.3.0
faiss-cpu
sentence-transformers>=2.2.2
rank_bm25
PyPDF2>=3.0.1
python-pptx>=0.6.21
python-docx
pandas
openpyxl
chromadb
pydantic
typing-extensions
python-dotenv